pyproject.toml
[project]
name = "nac-collector"
version = "0.1.0"
description = "A CLI tool to collect data from network infrastructure devices and systems."
license = { text = "MPL-2.0" }
readme = "README.md"
authors = [
    { name = "Kuba Mazurkiewicz", email = "kmazurki@cisco.com" },
]
maintainers = [
    { name = "Kuba Mazurkiewicz", email = "kmazurki@cisco.com" },
]
requires-python = ">=3.10"
keywords = ["network", "automation", "collection", "cisco", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
]
dependencies = [
    "typer>=0.17.4",
    "rich>=14.1.0",
    "gitpython>=3.1.45",
    "httpx>=0.28.1",
    "ruamel.yaml>=0.18.15",
    "tinydb>=4.8.2",
    "paramiko>=3.5.0",
]

[project.urls]
homepage = "https://github.com/netascode/nac-collector"
repository = "https://github.com/netascode/nac-collector"
documentation = "https://github.com/netascode/nac-collector"

[project.scripts]
nac-collector = "nac_collector.cli.main:app"

[project.optional-dependencies]
dev = [
    "bandit[toml]>=1.8.6",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.15.0",
    "ruff>=0.12.12",
    "types-paramiko>=3.5.0",
]

[tool.coverage.run]
source = ["nac_collector"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "@abstract", "except ImportError:"]
omit = ["*/__main__.py"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false
warn_unreachable = true
show_error_context = true

[tool.pytest.ini_options]
markers = ["unit", "integration"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["nac_collector"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert_used test

[tool.hatch.build.targets.wheel]
packages = ["nac_collector"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"